GBE: Import constantexpr lower pass from pNaCl
The idea is lower the constantExpr into an Instruction.

Fix the ptrtoInt and IntToPtr implementation, it simply maps to
a convert if type size not the same.

Fix a bitcast from integer to float issue. As we expand llvm::ConstantExpr
into llvm::Instruction. We will meet below situation.
  %10 = bitcast i32 1073741824 to float
  %11 = fcmp float %10 0.000000e+00
This will translated into GenIR:
  %100 = loadi S32 1073741824
  %101 = fcmp %100, 0.0f
In later instruction selection, we may directly getFloat() from %100

Signed-off-by: Ruiling Song <ruiling.song@intel.com>