libocl: define NULL to zero
using (void*)0 could not pass compilation in clang 3.6.
It will be treated as private address space pointer, if you compare
a global pointer with NULL, that is a private and global pointer
comparison, this is not allowed by OpenCL spec. But zero is allowed
as it is a pointer and integer comparison.

Detailed discussion, please read:
http://lists.cs.uiuc.edu/pipermail/cfe-dev/2015-February/041429.html

Signed-off-by: Ruiling Song <ruiling.song@intel.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>