GBE: unify element type before insertelement in legalize pass.
large integer type like i96 may be expanded to be low 64bit and high 32bit.
When it is cast to <i32 x 3>, we should first make the expanded type to be
of same type, here i32. insertelement could not insert element of different type.
Then we can do insertelement one by one to generate the <i32 x 3> vector.

This could fix the bug:
https://bugs.freedesktop.org/show_bug.cgi?id=89167

Signed-off-by: Ruiling Song <ruiling.song@intel.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>