Crash when hardware inaccessible
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=779213

Summary: On hardware where the Intel GPU is disabled, beignet was found
to assert-fail on load, taking the application down with it before it
can do anything (including checking for hardware via clGetDeviceIDs).

This fixes this crash, allowing existing error handling to return
CL_DEVICE_NOT_FOUND, and the application to then try other ICDs until
it finds the right one for the hardware.

Signed-off-by: Rebecca Palmer <rebecca_palmer@zoho.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>