runtime: fix a conformance bug in cl_get_kernel_arg_info.
Accordying to OpenCL 1.2 Rev 17:
"CL_KERNEL_ARG_TYPE_CONST is returned if the argument is a pointer and the referenced type is declared with the restrict or const qualifier. For
example, a kernel argument declared as global int const *x returns CL_KERNEL_ARG_TYPE_CONST but a kernel argument declared as global int *
const x does not."

So only need to return CL_KERNEL_ARG_TYPE_CONST for pointer arguments.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Tested-by: "Weng, Chuanbo" <chuanbo.weng@intel.com>