BUGFIX: Prohibit 'make package' from doing system install of ICD vendor file
As presently written, a 'make package' will attempt to INSTALL the
Beignet ICD loader to /etc/OpenCL/vendors whereas it should just
do a local install and then package the file. The proposed change instructs
CPack to include the `DESTDIR` variable when it calls `make install`, thus
directing the desination for the ICD loader to a local directory instead
of a system path.

Signed-off-by: Brian Kloppenborg <brian@arrayfire.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>