GBE: Use actual bti information to determine a pointer's addressspace.
Due to the private constant buffer support, it introduces private address
space mixed with constant address space some time. And more generic, one
constant address space may be located in private address space in LLVM IR
layer. Such as the following code:
  __kernel ...
  {
    const int2 foo[] = {{0, 1}, {2, 3}};
    int2 data = foo[get_global_id(0) % 2];
  }

The foo is in private address space but we finally will use __constant bti
to access it in Gen backend. The the above code will cause a assertion fail
in gen insturcion selection stage, because it generate a vector loading
instruction on a __constant buffer.

So we should use the actual BTI data to determine one pointer's address space
rather than get it from the LLVM IR layer.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Reviewed-by: "Yang, Rong R" <rong.r.yang@intel.com>