Kill the A0 cache in GenContext.
The a0 value cache in Gencontext can just hold the value
in compiling time, which may be different with the true
offset value in run time when the code generates the
backward jump. So just kill the cache of a0 and we will
use load vector instruction to optimize it lader.

Signed-off-by: Junyan He <junyan.he@linux.intel.com>
Reviewed-by: "Song, Ruiling" <ruiling.song@intel.com>