runtime: Need to separate atomic in L3 test and SLM test in self_test().
On HSW, if we use default 4.0.x kernel without the i915.enable_ppgtt=2
boot argument, then the atomic in L3 will not work and the SLM will not
work neither. We need to test atomic in L3 firstly, if it fails we need
to test SLM again. Otherwise, beignet will not give any error/warning
information for both atomic in L3 and SLM not working case.

v2:
shold set the atomic test result before the second round SLM test.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Reviewed-by: "Luo, Xionghu" <xionghu.luo@intel.com>