GBE/IMM: Temporarily allow integer type in getFloatValue()
The second argument of simd_shuffle is integer, which is not the
same as the instruction type. but when we are dealing with
BinaryInstruction in instruction selection, we would call
getFloatValue() to do some imm optimization. We should
implement a separate pattern for simd shuffle. after that,
I will revert this patch.

Signed-off-by: Ruiling Song <ruiling.song@intel.com>
Reviewed-by: "Yang, Rong R" <rong.r.yang@intel.com>