Backend: Implement FDIV64 on BDW.
According to the document, we use a set of instructions
to implement double type division.
Here, we specially take care of uniform register.  The
madm and invm instructions both work in align16 mode and
uniform register can not be used directly. So we need to
allocate a tmp register and mov the uniform src to it.
Align16 mode's channel mask is different from the align1,
So we can not directly write the result to the dst, and
need a tmp register to hold the result and MOV it to dst
later.

Signed-off-by: Junyan He <junyan.he@linux.intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>