Backend: Refine ConvertInstruction logic in insn_selection
The ConvertInstruction now need to handle a lot of special
cases instead of simple MOV. The judgement of native long
support, half support and reg restriction of long type and
the situation very complicated. The current code logic is
too verbose and hard to read. We now use sub routine functions
to make it clear and readable.

Signed-off-by: Junyan He <junyan.he@linux.intel.com>
Reviewed-by: "Yang, Rong R" <rong.r.yang@intel.com>