GBE: continue to refine interfering check.
More aggresive interfering check, even if both registers are in
Livein set or Liveout set, they are still possible not interfering
to each other.

v2:
Liveout interfering check need to take care those BBs which has only one
register defined.

For example:

BBn:
  ...
  MOV %r1, %src
  ...

Both %r1 and %r2 are in the BBn's liveout set, but %r2 is not defined or used
in BBn. The previous implementation ignore this BB which is incorrect. As %r1
was modified to a different value, it means %r1 could not be replaced with %r2
in this case.

v3:
Add comments and assertion to restrict the usage of interleve
check functions of DAG class.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Reviewed-by: Ruiling Song <ruiling.song@intel.com>