GBE: avoid vector registers when there is high register pressure.
If the reservedSpillRegs is not zero, it indicates we are in a
very high register pressure. Use register vector will likely
increase that pressure and will cause significant performance
problem which is much worse than use a short-live temporary
vector register with several additional MOVs.

So let's simply avoid use vector registers and just use a
temporary short-live-interval vector.

v2:
remove out-of-date comments.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Reviewed-by: Ruiling Song <ruiling.song@intel.com>