utests: fix compiler_fill_image_2d_array random bug
Use safer image write instead of map and memset. When
create image without data, we could not set pitch and
we don't know the pitch either. So use map and memset
the space is too dangerous if pitch is bigger than
w*sizeof(bpp), in this case the actually pitch is 512
but memset use pitch as 64*4=256. With only half space
set to 0, there will be undefined behavior when we want
to check the result for those space that we haven't set
to 0.

Signed-off-by: Pan Xiuli <xiuli.pan@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>