GBE:Add LLVM3.8 support
LLVM3.8 does not support iterator implicit cast to pointer, so add change
type for these uses. And due to the refine of ScalarEvolution, the using
namespace llvm::legacy is not enough to tell comilper which PassManager
to use, refine the PassManager declaration.
Now CallInst::hasStructRetAttr() will assert if there is no argments,
don't know if it is a llvm bug or our fault, just workaround it now.

V2:
Remove patch about CMakeList
Do not use macro to handle cast

Signed-off-by: Pan Xiuli <xiuli.pan@intel.com>
Reviewed-by: Ruiling Song <ruiling.song@intel.com>