enable byte gather for vload2/3/4/8(offset, char*) on SKL, BXT and BDW
currently, the unaligned bytes gather is implemented with readByteAsDWord
for a good performance, change back to native byte scattered read
for new GPU generations with better performance.

as for vload16(1/2/.../15, global char*), the native byte scattered
read on SKL/BXT is not good, so use the original method for vload16.

as for vload8(1/2/.../7, global char*), the native byte scattered
read on BDW is not good, so use the original method for vload8

V2: use UNPACK_BYTE to refine code
    disable vload8
    enable BDW support

Signed-off-by: Guo Yejun <yejun.guo@intel.com>
Reviewed-by: Ruiling Song <ruiling.song@intel.com>