Increase size for compile log output
The space to hold the output log from the compiler is currently
preallocated. The previous size (1000 chars) is too small when many
warnings and/or errors are present.

Enlarge the buffer to 1024*1024 chars, in order to allow up to 1024
errors, each taking 1024 bytes to report. This is still a rather
arbitrary choice, but should hopefully fit a wider range of cases.

(This fixes an issue reported on #opencl)

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>