Report build failures in backend to the build log
As noted at llvm_gen_backend:94, we currently lack a mechanism for
reporting failures in backend (beignet-managed) compiler passes to the
build log, and instead print the error to stderr and assert-fail.

This patch creates such a mechanism, and uses it for "function not found".
Please note that it has not had much testing yet.

Points for discussion/improvement:
-Does not currently print source location information.
-Are there other assertions that can be triggered by invalid input, and hence
should also use this?

Signed-off-by: Rebecca Palmer <rebecca_palmer@zoho.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>