GBE: Improve spill policy by considering use count.
In this patch, We use usecount/liverange as the spill cost.
And try to minimize the spill cost when do register spilling.
For registers inside loop, we will use 10**loop_depth to approximate
the use count.

When we spill for a scalar register, that is easy, just find
the candidate with least spill cost. But when spilling for a vector
of registers, we need to search in physical register space to make
enough room for the requested size, and then we will choose the group
of registers with least spill cost.

Note I change to always allocate for dst SelectionVector. As dst
SelectionVector normally don't introduce too much register pressure.
Allocating SelectionVector will greatly reduce MOVs. I think we
need to do further optimization for SelectionVector. We need to
do dynamic decision based on register pressure. If allowed, we should
try to allocate SelectionVector as much as possible, if register
pressure not allowed, then we need to try to add extra MOVs or even spilling.

v2:
fix assert in getSpillCost() for tmp register.

Signed-off-by: Ruiling Song <ruiling.song@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>