Backend: Add intel_sub_group_shuffle_down/up/xor with shuffle
We first get two result for the two src with shuffle then selection then
select the result with range condition, if we use if else with shuffle
the src may be influced by the if else prediction and the result may be
wrong.
Using the old shuffle with xor for shuffle_xor.
V2: Remove some useless old code

Signed-off-by: Pan Xiuli <xiuli.pan@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>