change the bahavior when writing to 3-component vector data types
Per OPenCL 1.2 spec 6.1.5: For 3-component vector data types,
the size of the data type is 4 * sizeof(component). The spec does
not explicitly say what the padding componenet will be, it should
be implementation-dependent, so, let's fill zero for the padding
componenet instead of just caring the 3 components, and the sends
instruction will be decreased (from 4 sends to 1 sends for short3*
case).

Signed-off-by: Guo Yejun <yejun.guo@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>