Apply image offset to read/write/map operations
If an image is created from a buffer, it need not start at the
beginning of the containing mappable object (especially for an image
with multiple planes, such as one created from VAAPI).  This fixes
the implementation of the clEnqueue{Read,Write,Map}Image functions to
take that offset into account.

Signed-off-by: Mark Thompson <sw@jkqxz.net>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>