toMB: use standard constant
0x followed by 15 `f` is 2^60-1, which is neither 2^64-1 (the maximum
value for an uint64_t) neither 2^44-1 (the value that would reside in
the upper 44 bits of a 64-bit numbers, which is what we want here).
It still work in the context, because the extra bits are just dropped
by the shift, but it's a rather odd constant _and_ it requires a type
specifier to avoid a warning (see previous commit).

Replace it by UINT64_MAX (2^64-1), which was probably the intended
value in the first place.

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>