Add command queue's enqueue thread.
According to Spec, event should be more suitable to implement
in async mode. We now add a thread to each command queue to
handle the event commands. The basic idea is:
1. If a command depends on other events which are not COMPLETED,
this command must be queued to that thread. Every event's status
change will notify the command queue, and give that thread a
chance to dequeue and run the enqueued commands.
2. For some BLOCK API, such as MapBuffer with BLOCK flag set, we
will wait for all the events in wait list ready and execute it
in sync mode, no event will be queued to that thread.
3. For NDRange like commands, because we want to gain the best
performance, we will check its wait list, if all are COMPLETED,
we SUBMIT that NDRange command, and set the event to SUBMUTTED
status. Event will also be queued to that thread, and that
thread will wait for it COMPLETED.

Signed-off-by: Junyan He <junyan.he@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>