Add a CMake option for toggling OCL ICD Loader compatibility
The new option allows anyone wishing to do so to explicitly disable
OCL ICD Loader support in Beignet, regardless of the presence or absence
of OCL ICD header files. This is particularly useful for people building
Beignet packages for distributions, as it avoids creating an implicit
dependency on the state of the build host. The new option defaults to ON
so the default behaviour of CMake configuration remains unchanged.

See also: https://bugs.freedesktop.org/show_bug.cgi?id=98885

Signed-off-by: Marek Szuba <Marek.Szuba@cern.ch>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>