Add cl_program define to runtime.
cl_program store the source and binary for OpenCL program building.
It will call the device->compiler_api to do the compile, build and
link job. The source of the program is only one while it has multi
binary results for each device. The built program has all kernels'
name available and can create kernel based on it.

Signed-off-by: Junyan He <junyan.he@intel.com>