Add cl_event define to runtime.
cl_event struct is important in runtime. Each clEnqueueXXX
like api will generate one cl_event, if the event can be executed
immediately, we exec it sync by calling the according device_api's
function. If not, we will store it in the command queue's worker
thead and exec it async when the events it depend on are all completed.

Signed-off-by: Junyan He <junyan.he@intel.com>